+++ R/add_p.R
@@ -58,7 +58,7 @@ add_p <- function(x, ...) {
 #' @examples
 #' # Example 1 ----------------------------------
 #' add_p_ex1 <-
-#'   trial[c("age", "grade", "response", "trt")] %>%
+#'   trial[c("age", "grade", "trt")] %>%
 #'   tbl_summary(by = trt) %>%
 #'   add_p()
 #'