+++ R/reexport.R
@@ -1,8 +1,3 @@
-# magrittr ---------------------------------------------------------------------
-#' @export
-#' @importFrom magrittr %>%
-magrittr::`%>%`
-
 # tibble -----------------------------------------------------------------------
 #' @export
 #' @importFrom tibble as_tibble
@@ -22,48 +17,4 @@ dplyr::vars
 #' @export
 dplyr::select
 
-# huxtable ---------------------------------------------------------------------
-#' @export
-#' @importFrom huxtable as_huxtable
-huxtable::as_huxtable
 
-# tidyselect -------------------------------------------------------------------
-#' @importFrom tidyselect starts_with
-#' @export
-tidyselect::starts_with
-
-#' @importFrom tidyselect ends_with
-#' @export
-tidyselect::ends_with
-
-#' @importFrom tidyselect contains
-#' @export
-tidyselect::contains
-
-#' @importFrom tidyselect matches
-#' @export
-tidyselect::matches
-
-#' @importFrom tidyselect num_range
-#' @export
-tidyselect::num_range
-
-#' @importFrom tidyselect all_of
-#' @export
-tidyselect::all_of
-
-#' @importFrom tidyselect any_of
-#' @export
-tidyselect::any_of
-
-#' @importFrom tidyselect everything
-#' @export
-tidyselect::everything
-
-#' @importFrom tidyselect last_col
-#' @export
-tidyselect::last_col
-
-#' @importFrom tidyselect one_of
-#' @export
-tidyselect::one_of