+++ R/vetted_models.R
@@ -4,7 +4,7 @@
 #' in [tbl_regression] and [tbl_uvregression].  If a model is passed to these
 #' functions and the model is not listed below and  a `tidy()` function is
 #' not specified in the `tidy_fun=` argument, the model object will be passed
-#' to [broom::tidy] or [broom.mixed::tidy].
+#' to `broom::tidy` or `broom.mixed::tidy`.
 #'
 #' \itemize{
 #'  \item{[stats::lm]}
@@ -19,7 +19,7 @@
 #' @name vetted_models
 #' @keywords internal
 #' @section model support:
-#' If [broom::tidy] or [broom.mixed::tidy] supports a class of model not listed
+#' If `broom::tidy` or `broom.mixed::tidy` supports a class of model not listed
 #' above, please submit a [GitHub Issue](https://github.com/ddsjoberg/gtsummary/issues).
 #' The model can be added to the list of vetted models. Unit tests will be
 #' put in place to ensure continued support for the model.