+++ tests/testthat/test-add_q.R
@@ -1,4 +1,5 @@
 context("test-add_q")
+testthat::skip_on_cran()
 
 table1 <- trial %>%
   tbl_summary(by = trt) %>%