+++ tests/testthat/test-bold_p.R
@@ -1,4 +1,5 @@
 context("test-bold_p")
+testthat::skip_on_cran()
 
 #### new comment
 test_that("no errors/warnings with standard use in tbl_summary() and add_p()", {