+++ tests/testthat/test-style_pvalue.R
@@ -1,4 +1,5 @@
 context("test-style_pvalue")
+testthat::skip_on_cran()
 
 test_that("no errors/warnings with standard use", {
   pvals <- c(